Modify: 리펙토링 중 - 데이터 초기화, View 함수 분리, 필요 없는 주석, 코드 삭제
